django_backend/requirements.txt
Django==4.2.10  # Web framework
djangorestframework==3.14.0  # REST API framework
django-cors-headers==4.3.1  # Handle CORS headers
dlib==19.24.2  # Machine learning & facial recognition
opencv-python==4.8.1.78  # Computer vision
numpy==1.26.3  # Numerical computations
Pillow==10.1.0  # Image processing
google-api-python-client==2.108.0  # Google API integration
google-auth==2.23.4  # Google authentication
google-auth-oauthlib==1.1.0  # OAuth authentication for Google services
reportlab==4.0.8  # PDF generation
djangorestframework-simplejwt==5.3.1  # JWT authentication support
python-decouple==3.8  # Manage environment variables
django-otp==1.3.0  # One-time password (OTP) support
pyotp==2.9.0  # OTP generation and verification
django-email-utils==1.0.0  # Email utilities (optional, can use Django's built-in email)
